Very very rough, just getting this up for discussion.

This implements fusion at graph build time for subsequent tensor operations (calls to `atop`). This can help reduce the graph size, reduce the overhead of graph building, and reduce unnecessary communication. We take the following approach.

- When possible, calls to `atop` return an `ATopArray` (a subclass of `Array`) that encodes the desired tensor operation without actually building the graph.
- When `ATopArray`s are passed to further calls to `atop`, their operations are extracted and fused into the subsequent `atop` output.
- When the `ATopArray.dask` property is requested, a single call to `atop` the encodes all the included operations is made. This could be cached (isn't currently).

`atop` operations are fuseable iff there are no "dummy indices". Note that no library code uses these, but there are tests for them and external users may be using them.

Todo:
- [ ] reorganize to not be a big mess
- [ ] tests tests tests
- [ ] fix `ATopArray` to work with mutable operations
- [x] improve runtime task interpreter
- [ ] figure out answers to the questions below

Fixes #2538