@jcrist 2017-08-15
**Open questions:** 

- Should computations ever be repeated?

Currently dask hashes all operations, meaning that

```python
z1 = x + y
z2 = -(x + y)
```

really operates as

```python
z1 = x + y
z2 = -z1
```

With fusion, this becomes a bit trickier. Sometimes it might be faster to repeat a computation if it reduces communication between tasks. Other times it might not. In testing I have witnessed this become a real problem, so we'll definitely want to handle this.

- Should fusion be broken off at a certain size?

As written now, fusion is greedy and will build up gigantic tasks if allowed to. These may take longer to run, resulting in poorer performance. This is coupled to the above, where we might want to break fusion off at a certain point. I'm not sure if this will actually be an issue, but it might be.

**A few ideas for how to fix the above:**

- Try not to repeat computations. We can still do this easily for the single output case (e.g. `x.compute()`), but we'd have to add another method to the dask collections to handle the multi-output cases (e.g. `dask.compute(x, y)`). I *think* this is probably what we want to do. We still couldn't avoid repeating in all cases though, as non-atop calls on intermediates would cause fusion to happen before compute:

```python
temp = x + y
out1 = z * temp
out2 = -out1
# could avoid recomputing all duplicate tasks here
dask.compute(out1, out2)
out3 = temp.sum()  # causes `temp` to be fused here
# couldn't avoid repeating the computation of `temp` here, as `.sum` causes
# `temp.dask` to be reified
dask.compute(out1, out2, out3)
```

- Stop fusion after a certain "width". After the subgraph reaches a certain width, convert the `ATopArray` to an `Array`. Other metrics might be total subtasks, or number of input arrays. This solves potential problem 2, but doesn't really address problem 1.

- Provide a `ATopArray.concrete()` method (with a better name) so the user can manually control fusion boundaries. Then they could write code like:

```python
out1 = ((x**2 + y**2) * z).concrete()  # this is one task
out2 = out1.dot(out.T)  # this is a separate task
```

and the barriers between tasks are manually defined.

---

I suspect that the best plan is to try the best to avoid repeating tasks, and also provide a `.concrete` method to convert `ATopArray`s into `Array`s to provide explicit boundaries.