@jcrist 2017-08-15
**Demo:**

```python
import time
import numpy as np
import dask.array as da
start = time.time()
x = da.random.random(size=(10000, 10000), chunks=(1000, 1000))
z = da.sin(x)**2 + da.cos(x)**2
res = z + 2 * z**2 + 3 * z**3
out = res.sum()
end = time.time()
print("Building graph took: %.2f s" % (end - start))
print("Graph has %d tasks" % len(out.dask))
start = time.time()
o = out.compute()
end = time.time()
print("Compute took: %.2f s" % (end - start))
```

**This PR:**
```
(dask) jcrist dask $ python test.py
Building graph took: 0.01 s
Graph has 239 tasks
Compute took: 1.88 s
```

**Master:**
```
(dask) jcrist dask $ python test.py
Building graph took: 0.02 s
Graph has 1339 tasks
Compute took: 2.08 s
```

Faster build time, smaller graph, slightly faster execution.

Note that due to the task repeating issue mentioned above, things like `res + res.T` are much slower, as the expression for `res` gets evaluated twice in each task - once for the original and once for the transpose.