@mrocklin 2017-08-15
Any objection if I add some tests to this?