@mrocklin 2017-08-15
Hope you don't mind.  I added a couple.  It's very cool to see this work. 

What do you think about putting the compound-atop-task into the task on it's own, something like 

    (apply_atop, my_atop_data, ('x', 0), ('y', 0))