@mrocklin 2017-08-15
This is cool

```python
In [1]: import dask.array as da

In [2]: %time x = da.ones((1000000, 1000000), chunks=(1000, 1000))
CPU times: user 1.25 s, sys: 32 ms, total: 1.29 s
Wall time: 1.29 s

In [3]: %time y = da.sin(x) ** 2 + da.cos(x) ** 2
CPU times: user 5.74 ms, sys: 53 µs, total: 5.79 ms
Wall time: 5.76 ms
```