@mrocklin 2017-08-15
> What do you think about putting the compound-atop-task into the task on it's own, something like

    (apply_atop, my_atop_data, ('x', 0), ('y', 0))

We might even add a little string repr for the `my_atop_data` object