@jcrist 2017-08-15
> What do you think about putting the compound-atop-task into the task on it's own

I think there are benefits to the callable object:
- Can make a nice repr for it (avoided for now)
- In the simple case of one task could return the original function, avoiding having two code paths
- Callables are cached when serialized to the distributed scheduler.