@jcrist 2017-09-12
> This seems to be stalled. Is there anything we can do now to push in some of this before it gets too stale?

I'm almost done with the interface PR, which this would rely on to remove some of the foot-guns. I'm fine with keeping this up to date with master, and plan to keep pushing on it, just been a hectic last couple of weeks.