@jakirkham 2017-09-27
FWIW am really excited about this. If there is something I can do to help push this forward (as I have time), please let me know.