@jakirkham 2017-10-31
By "interface PR" were you meaning PR ( https://github.com/dask/dask/pull/2748 )?